# Contracts/openapi.snapshot.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "BackendApi | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/clubs": {
      "get": {
        "tags": [
          "Clubs"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClubDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClubDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClubDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Clubs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClubRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClubRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClubRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClubDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClubDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClubDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clubs/{id}": {
      "put": {
        "tags": [
          "Clubs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClubRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClubRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClubRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClubDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClubDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClubDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Clubs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/clubs/{id}/home-site": {
      "put": {
        "tags": [
          "Clubs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClubHomeSiteRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClubHomeSiteRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClubHomeSiteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClubDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClubDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClubDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clubs/{id}/join": {
      "post": {
        "tags": [
          "Clubs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClubMemberDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClubMemberDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClubMemberDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clubs/{id}/leave": {
      "post": {
        "tags": [
          "Clubs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/clubs/{id}/members": {
      "get": {
        "tags": [
          "Clubs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClubMemberDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClubMemberDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClubMemberDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Clubs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddClubMemberRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddClubMemberRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddClubMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClubMemberDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClubMemberDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClubMemberDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clubs/{id}/members/{userId}": {
      "delete": {
        "tags": [
          "Clubs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/clubs/{id}/posts": {
      "post": {
        "tags": [
          "Clubs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClubPostRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClubPostRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClubPostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClubPostDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClubPostDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClubPostDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Clubs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClubPostDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClubPostDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClubPostDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clubs/{id}/materials": {
      "get": {
        "tags": [
          "Clubs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaterialDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaterialDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaterialDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddClubMemberRequest": {
        "required": [
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ClubDto": {
        "required": [
          "id",
          "name",
          "description",
          "facultyLeadUserId",
          "facultyLeadFullName",
          "studentInchargeUserId",
          "studentInchargeFullName",
          "homeSiteHtml",
          "memberCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "facultyLeadUserId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "facultyLeadFullName": {
            "type": [
              "null",
              "string"
            ]
          },
          "studentInchargeUserId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "studentInchargeFullName": {
            "type": [
              "null",
              "string"
            ]
          },
          "homeSiteHtml": {
            "type": [
              "null",
              "string"
            ]
          },
          "memberCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "ClubMemberDto": {
        "required": [
          "id",
          "userId",
          "userFullName",
          "joinedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "userFullName": {
            "type": "string"
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateClubPostRequest": {
        "required": [
          "content"
        ],
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
      "CreateClubRequest": {
        "required": [
          "name",
          "description",
          "facultyLeadUserId",
          "studentInchargeUserId"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "facultyLeadUserId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "studentInchargeUserId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
...
      "UpdateClubHomeSiteRequest": {
        "required": [
          "homeSiteHtml"
        ],
        "type": "object",
        "properties": {
          "homeSiteHtml": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "UpdateClubRequest": {
        "required": [
          "name",
          "description",
          "facultyLeadUserId",
          "studentInchargeUserId"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "facultyLeadUserId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "studentInchargeUserId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
...
    }
  },
...
}